Add new port devel/ocaml-cmdliner
Cmdliner is an OCaml module for the declarative definition of command
line interfaces.

It provides a simple and compositional mechanism to convert command line
arguments to OCaml values and pass them to your functions. The module
automatically handles syntax errors, help messages and UNIX man page
generation. It supports programs with single or multiple commands
(like darcs or git) and respects most of the POSIX and GNU conventions.

This module is required to update lang/js_of_ocaml