lang/clang3*, lang/clang-devel: set environment during build
The clang ports override the default build target, but the new targets
fail to set the environment like the default target does.  This patch
passes MAKE_ENV to environment.

It also combines compounds to a single make cmd and unsuppresses the
commands so the output shows on the build logs.

PR:		199500
Submitted by:	marino
Approved by:	brooks (maintainer)