devel/adacurses: work around FreeBSD 11 ncurses regression
The upgrade of ncurses on the -CURRENT version of FreeBSD resulted in an
regression (see PR) which results in multiple pkg-fallout mails to me
weekly.  Since the PR isn't moving despite periodic pinging, I am
forcing the ports version of ncurses to be used with FreeBSD 11 rather
than the base like FreeBSD 10 and earlier, and DragonFly.

While here, tweak a makefile install target that was emitting a non-fatal
error.

PR:	199109