lang/modula3: Upgrade 5.8.6 => pre-5.10.0
Modula3 hasn't had a proper release in close to 5 years.  However, the
project has moved to github and has been under slow continuous
development.  The internal version was recently switch to 5.10.0, but
this is not a release.

A new build script is used which simplifies the port makefile and even
the generation of new bootstrap compilers.  However, the original 5.8.6
bootstrap still works with some inline modifications.

This version still uses the gcc backend, but eventually it could be
switched to the new c-backend, something that DragonFly could use as
well.  The base has been moved from $LOCALBASE to $LOCALBASE/cm3, so
the "example" programs are back with the rest.