- Add py-python2-pythondialog 3.3.0
Python interface to the UNIX dialog utility and mostly-compatible programs
(Python 2 backport)

This version is a backport of pythondialog to Python 2. Unless you really have
to use Python 2, you should go to the pythondialog homepage and download the
reference implementation which, at the time of this writing (May 2015) and for
the forseeable future, is targeted at Python 3.

WWW: http://pythondialog.sourceforge.net/