multimedia/ffmpeg: Convert memalign-hack into an option
DragonFly already has an aligned malloc (posix_memalign) and the hack
in ffmpeg is tripping an assertion.  M.Dillon thinks the hack is stale.
In the spirit of no-op changes for FreeBSD, this changeset converts
the "--enable-memalign-hack" into an on-by-default option, and then
excludes the option for DragonFly.

It does incur a revbump though.
The last changes were made to multimedia/ffmpeg(0|26)
I'm doing this under "non-invasive DragonFly support" flag

P.S. I wouldn't be surprised if --enable-memalign-hack is wrong for
recent FreeBSD releases as well, so maybe the maintainers should take
a look at this new option.