games/gbottler: drop dependency on py-xml
py-xml is indeed required per gbottler docs, but api's it's rely on are
in standard python library since python 2.0, so just drop the dependency.
It starts fine, connects somewhere successfully etc without py-xml.

While here add NO_ARCH and limit python version to 2.x because of pygtk2.
It may worth to move it japanese/ because UI is single language (Japanese).