Update to a more recent development snapshot.
This prepares the terrain and makes it in sync with the upcoming Qt5 version of the port
that needs a more recent version than 0.9.2. There has been no official 0.9.3 yet, so all
distributions are packaging Launchpad snapshots of the port.

In addition to that:
- Rename the GENDOCS option to the more standard DOXYGEN.
- Simplify options handling.

Most of the work has been done in our experimental area51 repository by Tobias C. Berner
<tcberner@gmail.com>, thanks a lot!