[NEW] deskutils/py-taskw: Bindings for your taskwarrior database
This is a python API for the taskwarrior command line tool.

It contains two implementations: taskw.TaskWarriorShellout and
taskw.TaskWarriorDirect. The first implementation is the supported one
recommended by the upstream taskwarrior core project. It uses the task export
and task import commands to manipulate the task database.

The second implementation opens the task db file itself and directly
manipulates it. It exists for backwards compatibility, but should only
be used when necessary.

WWW: https://github.com/ralphbean/taskw