Varnish 3.0 is EoL
pecl-varnish supports Varnish 4.x as of version 1.2.0. However, there
are some caveats. The cChangelog for 1.2.0 mentions the following:

- Varnish 4.x compatibility, but
- VarnishStat and VarnishLog aren't supported when linking to Varnish
4.x anymore

I understand when Varnish 4.1 is released some old Varnish 3.x vmods
will be able to be updated. Perhaps this functionality will return also.
Unfortunately there is not much else that can be done.

This is better than the alternative which was to remove this port from
the ports tree.