net-p2p/btqueue: mark broken and deprecated
This port is runtime broken. btqueue scheduler starts successfully, and it
allows to add some torrents into the queue, but it just doesn't download
anything. While the same torrents were successfully fetched with deluge.
Tools like btshowmetainfo and btmakemetafile are working, but we have analogs
in the ports tree. The port is unmaintained and latest release is from 2006, so
mark it broken and deprecated, set expiration date to 2015-08-01.