security/amavisd-milter: Remove pre-everything check
The pre-everything target fails on systems that do not have sendmail
installed in base.  The logic may have worked in the past, but today
the existence check fails because the library that is checked is pulled
in later.

The check itself is aimed at the sysadmin that intentionally changes the
sendmail default to not install milter, but then tries to install this
port.  The worst case scenario the results from removing this check (even
if it worked) is that the amavisd-milter port breaks during the build.

PR:		200247
Submitted by:	marino
Approved by:	maintainer timeout (6 weeks)