net/py-soapy: set for deprecation
This library didn't get any updates for 14 years and it depends upon ancient
textproc/py-xml, that should be kicked from the ports tree. No ports in the tree
depending upon this one and we have dozen of modern and actively developed
alternatives:

 - net/py-soapy
 - net/py-suds
 - net/py-simplesoap etc.

So mark it deprecated and set expiration date to 2015-08-01 (to conform
expiration date of another py-xml consumer - net-p2p/btqueue, so we can kick
them both in the same day).

Discussed with: sbz (maintainer)