devel/py-phabricator: Support token-based authentication
Backport upstream pull request #23 [1] that adds support for new style
token-based authentication. This PR fixes issue #22 [2]. Additionally
modify the patch to add the token argument to the and of the argument
list as per upstream pull request comment.

This change fixes errors associated with not being able to find 'user'
or 'cert' configuration keys (in ~/.arcrc) keys when using token based
authentication.

This was reproducible with deskutils/py-bugwarrior

- Sort Makefile
- Sort USE_PYTHON values
- Canonicalize regression-test invocation

[1] https://github.com/disqus/python-phabricator/pull/23
[2] https://github.com/disqus/python-phabricator/issues/22

Approved by:		sbz (maintainer)
Differential Revision:	https://reviews.freebsd.org/D2983