Add 3 new uses
localbase: mostly a non user one which enforce the compiler related flags to
           lookup in localbase first to find libraries, designed to be used in
           other USES
libarchive: to be used each time one is using libarchive from ports. It is
            necessary to simplify work for porters dealing with different
            versions of libarchive from base and different ways libarchive can
            be linked in base (with libmd and/or libcrypto) only dealing with
            one libarchive over all the ports tree is easier
libedit:    enforce using libedit from ports for the same reasons as for
            libarchive

Same things will happen for other base libraries which collides with ports
version later.