Revert r394321: optimizing away MASTER_SITE_SUBDIR actually makes the port more fragile towards the next upgrades.  While it also does hide the moment of the next version update, we should generally try keeping ports fetchable at all times (and thus more robust).
This also reverts freefall login expansion in Makefile header: despite its
sanity and merit, it incidentally clashes with one of the PHB "rules" that
had recently plagued the Handbook.  Correct fix would be to fix the latter,
but for now this particular issue is causing too much drama on the lists.

Requested by:	mat