Mk/bsd.port.mk: Add PKG_CREATE_VERBOSE functionality
When PKG_CREATE_VERBOSE is set, the "-v" option will be added to
the arguments for PKG_CREATE.  The intended use is for poudriere and
other build monitors that can time out.  The verbose option of pkg
create will periodically emit output as the package is being created.
It would be set mk.conf (the DragonFly version of poudriere will set
it unconditionally during package building).

Approved by:	portmgr
Differential Revision:	https://reviews.freebsd.org/D3507