Require a C++11-compliant standard library and make it work with GCC < 4.9.
Despite ports r340913, the port still fails to build on FreeBSD 9.3 with
lang/gcc (GCC 4.8):

  libreify/src/parser.cc: In member function 'void Reify::Parser::parseProgram()':
  libreify/src/parser.cc:130:51: error: 'to_string' is not a member of 'std'

Apply a workaround for bug 193528 found in other ports to make that function
visible to GCC 4.8 (GCC 4.9 and later are fine).

Additionally, the port actually requires a C++11 standard library in
addition to a C++11-compliant compiler, so adjust the USES line and later
simplify the process of choosing which compiler to use.

Approved by:	vsevolod (maintainer)
Differential Revision:	https://reviews.freebsd.org/D3545