Upgrade to version 0.19.4.
This port seems to require the tzdata and combine applications,
but it actually needs httpoison too. If httpoison is not
installed during the build then tzdata will not be able to start
and a required ETS table will not be created, thus no timezone
info will ever be available and the build will fail.