gnatdroid-sysroot: Add Android API Level 21 (Lollipop)
The latest Android Native Development Kit (NDK) has API Level 21
in it (but not 20, nor 22 or the latest Level 23).  Add this option
to gnatdroid's sysroot port, and change the default API from Jelly Bean 1
(Level 16) to Kitkat (Level 19).

Bump gnatdroid's binutils and gnatdroid itself as a consequence of this
default change.  A new patch had to be added to lang/gcc-aux to handle
the CTYPE changes which haven't made to GCC yet.

Gnatdroid has been testing for building on all API's but not for
functionality beyond Level 16 due to lack of hardware.  I may soon
install an Android emulator to see if that will suffice.