www/py-django-contrib-comments: Fix *_DEPENDS
Update RUN_DEPENDS from file based back to PKGNAME based matching. Use
a glob pattern to match multiple version suffixed versions of the
Django ports, as `pkg info -g` is used to match them, excellent! [1]

This doesn't completely solve the problem for Django ports and their
dependencies, as there is still the issue of 'which' Django port to use
if:

 * The dependency is not yet satisfied (fresh installs), AND
 * The port that depends on Django supports multiple versions, like this
   one.

While I'm here:

- Mark architecture independent (NO_ARCH)
- Remove TEST_DEPENDS (no longer necessary)
- Rename test target since the framework supports all the things now.
- Update test target to include a default timezone, otherwise the
  test suite borks.

[1] Thanks antoine!