www/py-django-contrib-comments: Use Django 1.6 for RUN_DEPENDS
Default to using Django 1.6 (www/py-django16) because the current version
of www/py-django-mezzanine does not support anything later than 1.6, and
installation currently causes a conflict as installation of both django16
and django18 are attempted:

    py-django-mezzanine -> www/django16, www/py-django-contrib-comments
    py-django-contrib-comments -> www/django (1.8)

This fixes a new (py-django-mezzanine) failure in 2015Q4 reported by
pkg-fallout.

This is the least of all current evils, without having to update Mezzanine
to it's latest versions (*and* have to MFH that update). It has zero impact
on other ports as mezzanine is the only port that currently consumes/depends
on it.

While I'm here, switch back to file-based dependency declaration because
django* matches all django packages, and we can't anchor using $, resulting
in `pkg info "$1" potentially returning multiple matches.

Scripts/do-depends.sh should probably check for multiple results
and error out if there is.

Reviewed by:		antoine
MFH:			2015Q4 (pre-approved: antoine)
Differential Revision:	https://reviews.freebsd.org/D4047