Switch back to tracking the upstream trunk instead of OpenOffice 4.1.2 release candidates.  Upgrade to SVN revision r1712473.
Add an option to build and install the PDF Import extension and enable
it by default.

Patch the extensions.lst file so that it points to the locations
of the extension files rather than passing that information as
a configure option.  Add a patch to enable enable the use of file://
in extensions.lst for local integrated extensions as is documented in
the comments in this file.

Nuke the patches that have been committed upstream.

Don't try to run the create_tree.sh script with "sh -c" since this
script is not marked executable.

MFH:		2015Q4