databases/firebird25-(client|server): convert to USES=libedit
The -server port had to be bumped because libedit was listed as a
BUILD_DEPENDS but libedit is linked dynamically so it has to be a
LIB_DEPENDS.  The error was masked by the fact the client is usually
present which ensures libedit.so is available.

While here, remove obsolete CONFLICT* definitions and unmask a bunch
of installation commands.