Update cloudabi-utils to version 0.7.
This new version of cloudabi-utils now installs a library called
libcloudabi. This library contains native ports of some of the APIs that
are currently provided as part of CloudABI's runtime. These used to be
built into cloudabi-run itself, but are now exposed as well.

As cloudabi-utils is now becoming more complex, it now uses CMake as its
build system. We can now remove our custom build targets. We do need to
make sure to override CMAKE_INSTALL_MANDIR, as it is set to share/man by
default. I'm slightly amazed that this isn't done by default.

Reviewed by:	beat
Differential Revision:	https://reviews.freebsd.org/D4176