audio/murmur: Fix segfault on FreeBSD 9.x, use base OpenSSL
FreeBSD 9.x was segfaulting with the update to 1.2.10 due to the
change to build against the ports OpenSSL. The problem was that while
murmur was linked against ports OpenSSL, it was also linked to QT which
pulled in the base system OpenSSL. FreeBSD 10.x and head did not
segfault due to pure luck.

While here, require compiler:c++11-lang. This is definitely required for
the upcoming 1.3.0 release, but also I want to avoid and weirdness on
FreeBSD 9.x which was building murmur with gcc but required a
clang-built boost at build time.