Update PyQt4 to 4.11.4, SIP to 4.17 and QScintilla2 to 2.9.1.
Bring in some long overdue updates, some of which are required for us to
later land the PyQt5 ports.

One big change with this update is that the PyQt4 ports now install their
.sip files into share/py-sip/PyQt4 instead of share/py-sip. This way we do
not end up with directories like share/py-sip/QtCore, which are especially
confusing once PyQt5 lands and starts installing files with the same names.

Other noteworthy items:
- PORTREVISION has been bumped on ports depending on devel/qscintilla2
  because libqscintilla2.so's SOVERSION has changed.
- graphics/seexpr has been converted to USE_PYQT, as the file it used to
  define a build-time dependency on x11-toolkits/py-qt4-gui has moved.

Once again, big thanks to Tobias Berner <tcberner@gmail.com> and Guido Falsi
(madpilot@) for their initial work on these ports as part of the effort to
land PyQt5 into the tree (see D2910 in Phabricator for an earlier version of
the PyQt5 patch set).

PR:		205143