lang/fpc-units, devel/fpc-bfd: Use binutils instead of ancient libbfd
Currently, lang/fpc-bfd can't be installed if devel/binutils is already
on the system.  Given that this is a requirement for lang/gcc and other
big patches, chances are there will be a conflict.  Resolve the problem
by making devel/fpc-bfd depend on devel/binutils instead.  This also
allows the BFD option on FreeBSD9 to be re-enabled.