Start fetching from GitHub and unmark BROKEN.
The original MASTER_SITE is gone, and the project appears to be maintained in
GitHub these days. diff -upr between the old zip file from
distcache.FreeBSD.org and the new tarball from GitHub only shows that the .svn/
directory is no longer shipped, and the $Id$ tags in the files are not expanded
anymore.