devel/gitinspector: support build with LANG=C (fix from upstream)
Gitinspect is now 4 releases behind.  The problem where gitinspector will
not build when LANG is set to "C" in the environment has been fixed for
14 months already.  Apply the patch from the git repository upstream to
fix (due to luck in name, it applies before patch-aa which affects the
same localization file).  I recommend that this port be upgraded to a new
release though.

Approved by:	just fix it