MFH: r407068
In the copy of mednafen included in libretro-cores, replace a named
label in inline assembly in an inline function with a local label.
This prevents "invalid symbol redefinition" errors when the function is
inlined multiple times, for example within an unrolled loop.

Approved by:	yuri@rawbw.com (maintainer)
PR:		206542

Approved by:	ports-secteam (with hat)