Remove CONFLICTS with older qt5-designer versions.
It is causing portmaster to fail with an error like this:
  ===>>> x11-toolkits/qt5-uiplugin seems to depend on devel/qt5-designer
         which looks like a dependency loop

which is not really what's happening here; the error comes from the fact that
qt5-designer 5.4.1 is being upgraded to 5.5.1, qt5-uiplugin is added as a
dependency that must be built first, but then it's marked as CONFLICTing with
the installed qt5-designer version.

Since qt5-designer < 5.5.0 does not actually install any of the uiplugin files
(since it was not an independent module before), we can just remove the
CONFLICTS line.

Thanks to Walter Schwarzenfeld (w.schwarzenfeld@aon.at) for the report.

PR:		206435