Document webkit CVE-2014-1748.
If people look at the announcement, CVE-2014-3192 is already fixed. This
CVE was against chromium, and the same code in 2.4.9 is in webkit trunk
so I assume it already fixed.

CVE-2013-6663 is for webkit < 2.4.0, and the rest of the CVE's are for
apple products without any attached patches.

PR:		205683
Obtained from:	http://webkitgtk.org/security/WSA-2015-0002.html