New port: editors/with-editor
The library with-editor makes it easy to use the Emacsclient as the
$EDITOR of child processes, making sure they know how to call
home. For remote processes a substitute is provided, which
communicates with Emacs on standard output instead of using a socket
as the Emacsclient does.

WWW: https://github.com/magit/with-editor

PR:		206767
Submitted by:	Yasuhiro KIMURA <yasu@utahime.org>