x11/kde4-workspace: fix OpenGL tests with nVidia's libGL
nVidia's libGL causes a segfault in ld-elf if the application that is linking
to libGL also links to another library (ordered before linking to libGL) and
the second library links to libthr.  For example:

kwin_opengl_test
 -> libXft
   -> libthr
 -> libGL

Two workarounds are known:
 1) Change the linking order to have libGL linked first, or
 2) Also link to libthr in the (base) application.

This patch implements the latter fix.

PR:	205149, 206899
MFH:	2016Q1