This is the adaptation of PlayOnLinux to FreeBSD. The purpose of this program is to simplify installation of MS Windows software to run under the Windows emulator "Wine".
The name indicates, that the program was initially developed to support
easy installation of Windows games on Linux. The range of supported
applications has long been extended to cover a wide range of windows
applications including Microsoft Office.

PlayOnFreeBSD behaves significantly different than PlayOnLinux in a
number of aspects, e.g. which version of Wine to use (only a single
version is normally installed on FreeBSD, but PlayOnLinux tries to run
each Windows application in a Wine environment that has previously been
reported to give good results and installs missing Wine versions in the
user's home directory).

Reviewed by:	antoine