lang/gcc6-aux: install gnatdroid/dejagnu support
Change names of gnatdroid-support dejagnu exp files so that gcc6-aux
can co-exist with gcc-aux.  The gnatdroid compilers will be rebased
on gcc6-aux in the future.  Technically the resulting package changes
and thus requires a bump, but nothing uses these two new files yet,
so I'm not bumping it.