Merge cloudabi-clang with cloudabi-toolchain and install more useful tools.
Now that the previous changes to the cloudabi-clang package brought in
more LLVM tools, the package actually became a misnomer. Most of the
tools that it installed are not part of Clang, but of LLVM instead. My
plans of extending it to also install wrappers around man(1) and
pkgconf(1) would make the naming even worse.

The only reason that the cloudabi-clang package was created in the first
place, was because we still had some of the core CloudABI libraries part
of FreeBSD Ports. Now that these libraries are part of their own package
collection, it makes more sense to squash this package together with
cloudabi-toolchain.

The nice thing about installing a wrapper around pkgconf(1) named
${gnu_triple}-pkg-config is that Autoconf now automatically picks it up.
Running ./configure --host=${gnu_triple} is enough to make it detect the
cross compiled libraries it depends on.

Reviewed by:	bapt
Differential Revision:	https://reviews.freebsd.org/D5697