New port: "Yet Another APRS Client" - A Java-based graphical APRS client.
YAAC can be used as a stand-alone APRS client, an APRS RF-Internet gateway
(I-Gate), or as a AX.25 digipeater. It supports 16 different data reporting
views as delivered, and can be extended with user-written "plugins" to add
more functionality. YAAC's map rendering does not depend on Internet
connectivity, so it can be used wherever a portable computer and radio/TNC
can be located.