Try harder to find svn
Don't assume svn is installed, but use it if it's available, falling back to
svnlite if not (or using svn from base in the case where WITH_SVN was set in
base). This matches the logic in base for finding svn.

PR:		207479
Discussed with:	gjb
Reported by:	Torsten Zühlsdorff <ports@toco-domains.de>