Cache ac_cv_sys_long_file_names conftest result (always yes)
There are a few ports that check for long file name support during
configuration, but the conftest actually modifies $LOCALBASE/lib when
the port is built by the root user.  By caching the result (yes), the
test is avoiding and so is the consequential file system violation.
(The synth test option can detect this FS violation.)

Differential Revision:	https://reviews.freebsd.org/D6160
Approved by:		portmgr (bapt@, bdrewery@)