Fix build of openoffice-4 and openoffice-devel with libc++ 3.8.0.
Openffice defines its own overloaded variant of round(), which is
only different in return type from the 'real' round() from <math.h>,
so it is ambiguous (in the C++ lookup sense).  Fix this by renaming
the custom round() to round_(), and using a define to minimize
needed changes in the code.

PR:		209588
Submitted by:	dim