cad/ghdl: build it with gcc6-aux, not gcc5-aux
I confirmed that ghdl builds with the default-for-Ada gcc6-aux compiler
on FreeBSD.  It was building on DragonFly, but DragonFly uses the LLVM
backend while on FreeBSD ghdl defaults to gcc back, so both options
needed to be tested.  It's debatable that a bump is needed because ghdl
is self-contained, but let's do it anyway so they'll be immediate
feedback on any runtime issues this may cause.