www/impresscms: regenerate pkg-plist, bump
This ports appears to have been broken ever since it was provided stage
support incorrectly.  All the @dirrm entries were converted to @dir,
and not all of those were even pointing at directories.  This regenerates
the plist with "makeplist" so that it passes build tests.

Note: version 1.3.8 is out now in case somebody wants to upgrade this.